requirements.txt
mysql-connector-python==8.0.33
matplotlib==3.8.0
bcrypt==4.0.1
tkcalendar
